A bunch of fix for web and nl web server
