+++ pytest.ini
@@ -1,0 +1,4 @@
+[pytest]
+log_cli=true
+log_level=INFO
+log_file=log.txt
